+++ src/models/v2v3/contracts.ts
@@ -23,3 +23,22 @@ export enum V2V3ContractName {
 }
 
 export type V2V3Contracts = Record<V2V3ContractName, Contract>
+
+/**
+ * Single source of truth for supported terminal versions.
+ *
+ * DEV NOTE:
+ * To support a new payment terminal:
+ * 1. Add it to V2V3ContractName
+ * 2. Add it to this array
+ * 3. Add support for it in any other transactions that use it.
+ */
+export const SUPPORTED_PAYMENT_TERMINALS = [
+  V2V3ContractName.JBETHPaymentTerminal,
+  V2V3ContractName.JBETHPaymentTerminal3_1,
+  V2V3ContractName.JBETHPaymentTerminal3_1_1,
+  V2V3ContractName.JBETHPaymentTerminal3_1_2,
+] as const
+
+export type PaymentTerminalVersion =
+  (typeof SUPPORTED_PAYMENT_TERMINALS)[number]
